Попап да/нет и страничка выбора персонажа;

+++ BRIX.Mobile/ViewModel/Characters/CharacterListPageVM.cs
@@ -2,10 +2,13 @@
 using BRIX.Mobile.Services;
 using BRIX.Mobile.Services.Navigation;
 using BRIX.Mobile.View.Characters;
+using BRIX.Mobile.View.Popups;
 using BRIX.Mobile.ViewModel.Base;
+using BRIX.Mobile.ViewModel.Popups;
 using CommunityToolkit.Mvvm.ComponentModel;
 using CommunityToolkit.Mvvm.Input;
 using System.Collections.ObjectModel;
+using BRIX.Mobile.Resources.Localizations;
 using CharacterBM = BRIX.Library.Characters.Character;
 
 namespace BRIX.Mobile.ViewModel.Characters
@@ -13,10 +16,12 @@ namespace BRIX.Mobile.ViewModel.Characters
     public partial class CharacterListPageVM : ViewModelBase
     {
         private readonly ICharacterService _characterService;
+        private readonly ILocalizationResourceManager _localization;
 
-        public CharacterListPageVM(ICharacterService characterService)
+        public CharacterListPageVM(ICharacterService characterService, ILocalizationResourceManager localization)
         {
             _characterService = characterService;
+            _localization = localization;
         }
 
         [ObservableProperty]
@@ -54,9 +59,20 @@ namespace BRIX.Mobile.ViewModel.Characters
         [RelayCommand]
         private async Task Remove(CharacterModel character)
         {
-            //TODO: вывести предупредительное сообщение
-            await _characterService.RemoveAsync(character.Id);
-            Characters.Remove(character);
+            QuestionPopupResult result = await ShowPopupAsync<QuestionPopup, QuestionPopupResult, QuestionPopupParameters>(
+                new QuestionPopupParameters(
+                    title: _localization[LocalizationKeys.Warning].ToString(),
+                    message: _localization[LocalizationKeys.DeleteCharacterQuestion].ToString(),
+                    yesText: _localization[LocalizationKeys.Yes].ToString(),
+                    noText: _localization[LocalizationKeys.No].ToString()
+                )
+            );
+
+            if (result?.Answer == EQuestionPopupResult.Yes)
+            {
+                await _characterService.RemoveAsync(character.Id);
+                Characters.Remove(character);
+            }
         }
 
         [RelayCommand]
